LayoutTests/imported/w3c/web-platform-tests/dom/events/non-cancelable-when-passive/passive-mousewheel-event-listener-on-div-expected.txt


FAIL passive mousewheel event listener on div assert_equals: expected false but got true
